Move `AppList` to `containers`
