+++ lib/app/containers/App.js
@@ -6,7 +6,7 @@ const Miss = require('react-router/Miss').default
 const Header = require('../components/Header')
 const Footer = require('../components/Footer')
 const Drawer = require('../components/Drawer')
-const AppList = require('../components/AppList')
+const AppList = require('./AppList')
 
 const { filterApps } = require('../actions')
