+++ lib/app/containers/AppList.js
@@ -2,12 +2,12 @@ const debounce = require('lodash.debounce')
 const React = require('react')
 const { connect } = require('react-redux')
 
-const { fetchApps, filterApps, searchApps } = require('../../actions')
+const { fetchApps, filterApps, searchApps } = require('../actions')
 
-const AppCard = require('./AppCard')
-const CountriesModal = require('./CountriesModal')
-const TechnologiesModal = require('./TechnologiesModal')
-const TagsModal = require('./TagsModal')
+const AppCard = require('../components/AppList/AppCard')
+const CountriesModal = require('../components/AppList/CountriesModal')
+const TechnologiesModal = require('../components/AppList/TechnologiesModal')
+const TagsModal = require('../components/AppList/TagsModal')
 
 const { PropTypes, Component } = React
